BUILD: publish to default GitHub pages address only when releasing the latest stable version (#204)
* REFACTOR: move listing of versions to separate Versions class
* deploy docs to main path only if releasing the latest stable version
* use version parsing for listing versions in make_base.py
* update version to 1.1.0rc1
* rename non-rc to stable
* replace classmethod pattern with singleton pattern
* CI/CD: Clarify log messages during doc build

+++ sphinx/base/make_base.py
@@ -13,9 +13,11 @@ import sys
 from abc import ABCMeta, abstractmethod
 from glob import glob
 from tempfile import TemporaryDirectory
-from typing import Dict, Iterable, List, Set, Tuple, Type
+from typing import Dict, Iterable, List, Set, Tuple
 
 from packaging import version as pkg_version
+
+from pytools.meta import SingletonMeta, compose_meta
 
 cwd = os.getcwd()
 
@@ -50,38 +52,30 @@ DIR_ALL_DOCS_VERSIONS = os.path.join(DIR_SPHINX_BUILD, "docs-version")
 ENV_PYTHON_PATH = "PYTHONPATH"
 
 
-class Command(metaclass=ABCMeta):
+class Command(metaclass=compose_meta(ABCMeta, SingletonMeta)):
     """ Defines an available command that can be launched from this module."""
 
     __RE_CAMEL_TO_SNAKE = re.compile(r"(?<!^)(?=[A-Z])")
 
-    @classmethod
-    def get_name(cls) -> str:
-        try:
-            return cls.__name
-        except AttributeError:
-            cls.__name = cls.__RE_CAMEL_TO_SNAKE.sub("_", cls.__name__).lower()
-            return cls.__name
+    def __init__(self) -> None:
+        self.name = self.__RE_CAMEL_TO_SNAKE.sub("_", type(self).__name__).lower()
 
-    @classmethod
     @abstractmethod
-    def get_description(cls) -> str:
+    def get_description(self) -> str:
         pass
 
-    @classmethod
-    def get_dependencies(cls) -> Tuple[Type["Command"], ...]:
-        return ()
+    @abstractmethod
+    def get_dependencies(self) -> Tuple["Command", ...]:
+        pass
 
-    @classmethod
-    def get_prerequisites(cls) -> Iterable[Type["Command"]]:
-        dependencies_extended: List[Type["Command"]] = []
+    def get_prerequisites(self) -> Iterable["Command"]:
+        dependencies_extended: List[Command] = []
 
-        for dependency in cls.get_dependencies():
+        for dependency in self.get_dependencies():
             dependencies_inherited = dependency.get_dependencies()
-            if cls in dependencies_inherited:
+            if self in dependencies_inherited:
                 raise ValueError(
-                    f"circular dependency: {dependency.get_name()} "
-                    f"depends on {cls.get_name()}"
+                    f"circular dependency: {dependency.name} " f"depends on {self.name}"
                 )
             dependencies_extended.extend(
                 dependency
@@ -92,14 +86,12 @@ class Command(metaclass=ABCMeta):
 
         return dependencies_extended
 
-    @classmethod
-    def run(cls) -> None:
-        print(f"Running command {cls.get_name()} – {cls.get_description()}")
-        cls._run()
+    def run(self) -> None:
+        print(f"Running command {self.name} – {self.get_description()}")
+        self._run()
 
-    @classmethod
     @abstractmethod
-    def _run(cls) -> None:
+    def _run(self) -> None:
         pass
 
 
@@ -109,12 +101,13 @@ class Command(metaclass=ABCMeta):
 
 
 class Clean(Command):
-    @classmethod
-    def get_description(cls) -> str:
+    def get_description(self) -> str:
         return "remove Sphinx build output"
 
-    @classmethod
-    def _run(cls) -> None:
+    def get_dependencies(self) -> Tuple[Command, ...]:
+        return ()
+
+    def _run(self) -> None:
         if os.path.exists(DIR_SPHINX_BUILD):
             shutil.rmtree(path=DIR_SPHINX_BUILD)
         if os.path.exists(DIR_SPHINX_API_GENERATED):
@@ -124,17 +117,13 @@ class Clean(Command):
 
 
 class ApiDoc(Command):
-    @classmethod
-    def get_description(cls) -> str:
+    def get_description(self) -> str:
         return "generate Sphinx API documentation from sources"
 
-    @classmethod
-    def get_dependencies(cls) -> Tuple[Type["Command"], ...]:
-        # noinspection PyRedundantParentheses
-        return (Clean,)
+    def get_dependencies(self) -> Tuple[Command, ...]:
+        return (Clean(),)
 
-    @classmethod
-    def _run(cls) -> None:
+    def _run(self) -> None:
         packages = [
             package for package in os.listdir(DIR_PACKAGE_SRC) if package[:1].isalnum()
         ]
@@ -173,17 +162,13 @@ class ApiDoc(Command):
 
 
 class GettingStartedDoc(Command):
-    @classmethod
-    def get_description(cls) -> str:
+    def get_description(self) -> str:
         return "generate getting started documentation from sources"
 
-    @classmethod
-    def get_dependencies(cls) -> Tuple[Type["Command"], ...]:
-        # noinspection PyRedundantParentheses
-        return (Clean,)
+    def get_dependencies(self) -> Tuple[Command, ...]:
+        return (Clean(),)
 
-    @classmethod
-    def _run(cls) -> None:
+    def _run(self) -> None:
 
         # make dir if it does not exist
         os.makedirs(DIR_SPHINX_GSTART_GENERATED, exist_ok=True)
@@ -219,44 +204,19 @@ class GettingStartedDoc(Command):
 
 
 class FetchPkgVersions(Command):
-    @classmethod
-    def get_description(cls) -> str:
+    def get_description(self) -> str:
         return "fetch available package versions with docs"
 
-    @classmethod
-    def get_dependencies(cls) -> Tuple[Type["Command"], ...]:
+    def get_dependencies(self) -> Tuple[Command, ...]:
         return ()
 
-    @classmethod
-    def _run(cls) -> None:
-        os.makedirs(DIR_SPHINX_BUILD, exist_ok=True)
-        start_from_version_tag = "1.0.1"
-        sp = subprocess.run(
-            args='git tag -l "*.*.*"', shell=True, check=True, stdout=subprocess.PIPE
-        )
-        version_tags = sp.stdout.decode("UTF-8").split("\n")
-        version_tags = [
-            vt for vt in version_tags if vt != "" and vt >= start_from_version_tag
-        ]
-
-        # add version currently build into version_tags
-        version_built = get_package_version()
-
-        if version_built not in version_tags:
-            version_tags.append(version_built)
-
-        version_tags.sort()
-        version_tags.reverse()
-        version_tags_non_rc = [vt for vt in version_tags if "rc" not in vt]
-        latest_non_rc_version = version_tags_non_rc[0]
-
-        print("Found the following version tags: ", version_tags)
-        print("Latest non-RC version: ", latest_non_rc_version)
+    def _run(self) -> None:
+        versions = Versions()
 
         version_data = {
-            "current": latest_non_rc_version,
-            "non_rc": version_tags_non_rc,
-            "all": version_tags,
+            "current": str(versions.latest_stable_version),
+            "non_rc": list(map(str, versions.version_tags_stable)),
+            "all": list(map(str, versions.version_tags)),
         }
 
         version_data_as_js = (
@@ -270,34 +230,43 @@ class FetchPkgVersions(Command):
 
 
 class PrepareDocsDeployment(Command):
-    @classmethod
-    def get_description(cls) -> str:
+    def get_description(self) -> str:
         return "update versions of rendered documentation"
 
-    @classmethod
-    def get_dependencies(cls) -> Tuple[Type["Command"], ...]:
+    def get_dependencies(self) -> Tuple[Command, ...]:
         return ()
 
-    @classmethod
-    def _run(cls) -> None:
-        assert is_azure_build(), "Only implemented for Azure Pipelines"
+    def _run(self) -> None:
+        if not is_azure_build():
+            raise RuntimeError("only implemented for Azure Pipelines")
+
+        # get the current version of the package
+        current_version: pkg_version.Version = get_package_version()
+
+        # copy new the docs version to the deployment path
+        if current_version == Versions().latest_stable_version:
+            # only copy to deployment path if our version is the latest stable release
+
+            # remove docs build currently deployed, except for the docs versions folder
+            if os.path.exists(os.path.join(DIR_DOCS, "docs-version")):
+                with TemporaryDirectory() as DIR_TMP:
+                    shutil.move(src=os.path.join(DIR_DOCS, "docs-version"), dst=DIR_TMP)
+                    shutil.rmtree(path=DIR_DOCS)
+                    os.makedirs(DIR_DOCS)
+                    shutil.move(src=os.path.join(DIR_TMP, "docs-version"), dst=DIR_DOCS)
+            else:
+                os.makedirs(DIR_DOCS, exist_ok=True)
+
+            # copy new docs version to deployment path
+            shutil.copytree(src=DIR_SPHINX_BUILD_HTML, dst=DIR_DOCS, dirs_exist_ok=True)
+
         # get current version of package in the form of folder/URL name (e.g., "1-0-0")
-        current_version = version_string_to_url(get_package_version())
-        # remove docs build currently deployed, except for the docs versions folder
-        if os.path.exists(os.path.join(DIR_DOCS, "docs-version")):
-            with TemporaryDirectory() as DIR_TMP:
-                shutil.move(src=os.path.join(DIR_DOCS, "docs-version"), dst=DIR_TMP)
-                shutil.rmtree(path=DIR_DOCS)
-                os.makedirs(DIR_DOCS)
-                shutil.move(src=os.path.join(DIR_TMP, "docs-version"), dst=DIR_DOCS)
-
-        # copy new docs version to deployment path
-        os.makedirs(DIR_DOCS, exist_ok=True)
-        shutil.copytree(src=DIR_SPHINX_BUILD_HTML, dst=DIR_DOCS, dirs_exist_ok=True)
-
+        current_version_path = os.path.join(
+            DIR_DOCS, "docs-version", version_string_to_url(current_version)
+        )
         # update latest version in docs history
-        if os.path.exists(os.path.join(DIR_DOCS, "docs-version", current_version)):
-            shutil.rmtree(path=os.path.join(DIR_DOCS, "docs-version", current_version))
+        if os.path.exists(current_version_path):
+            shutil.rmtree(path=current_version_path)
         shutil.copytree(
             src=DIR_ALL_DOCS_VERSIONS,
             dst=os.path.join(DIR_DOCS, "docs-version"),
@@ -322,16 +291,13 @@ class PrepareDocsDeployment(Command):
 
 
 class Html(Command):
-    @classmethod
-    def get_description(cls) -> str:
+    def get_description(self) -> str:
         return "build Sphinx docs as HTML"
 
-    @classmethod
-    def get_dependencies(cls) -> Tuple[Type["Command"], ...]:
-        return Clean, FetchPkgVersions, ApiDoc, GettingStartedDoc
+    def get_dependencies(self) -> Tuple[Command, ...]:
+        return Clean(), FetchPkgVersions(), ApiDoc(), GettingStartedDoc()
 
-    @classmethod
-    def _run(cls) -> None:
+    def _run(self) -> None:
 
         check_sphinx_version()
 
@@ -353,7 +319,7 @@ class Html(Command):
         sys.path.append(DIR_MAKE_BASE)
 
         # create copy of this build for the docs archive
-        version_built = get_package_version()
+        version_built: pkg_version.Version = get_package_version()
         dir_path_this_build = os.path.join(
             DIR_ALL_DOCS_VERSIONS, version_string_to_url(version_built)
         )
@@ -376,13 +342,58 @@ class Html(Command):
 
 
 class Help(Command):
-    @classmethod
-    def get_description(cls) -> str:
+    def get_description(self) -> str:
         return "print this help message"
 
-    @classmethod
-    def _run(cls) -> None:
+    def get_dependencies(self) -> Tuple[Command, ...]:
+        return ()
+
+    def _run(self) -> None:
         print_usage()
+
+
+class Versions(metaclass=SingletonMeta):
+    """
+    Helper class that lists all versions that have already been released.
+    """
+
+    INITIAL_VERSION_TAG = pkg_version.parse("1.0.1")
+
+    def __init__(self) -> None:
+        os.makedirs(DIR_SPHINX_BUILD, exist_ok=True)
+        start_from_version_tag: pkg_version.Version = Versions.INITIAL_VERSION_TAG
+        sp = subprocess.run(
+            args='git tag -l "*.*.*"', shell=True, check=True, stdout=subprocess.PIPE
+        )
+        version_tags: List[pkg_version.Version] = [
+            version_tag
+            for version_tag in (
+                pkg_version.parse(version_string)
+                for version_string in sp.stdout.decode("UTF-8").split("\n")
+                if version_string
+            )
+            if version_tag >= start_from_version_tag
+        ]
+
+        # add version currently build into version_tags
+        version_built: pkg_version.Version = get_package_version()
+
+        if version_built not in version_tags:
+            version_tags.append(version_built)
+
+        version_tags.sort()
+        version_tags.reverse()
+        version_tags_stable: List[pkg_version.Version] = [
+            vt for vt in version_tags if not (vt.is_prerelease or vt.is_devrelease)
+        ]
+        latest_stable_version: pkg_version.Version = version_tags_stable[0]
+
+        print(f"Found versions: {', '.join(map(str, version_tags))}")
+        print("Latest stable version: ", latest_stable_version)
+
+        self.version_tags = version_tags
+        self.version_tags_stable = version_tags_stable
+        self.latest_stable_version = latest_stable_version
 
 
 def make(*, modules: List[str]) -> None:
@@ -411,11 +422,11 @@ def make(*, modules: List[str]) -> None:
     os.environ[ENV_PYTHON_PATH] = os.pathsep.join(module_paths)
 
     # run all given commands:
-    executed_commands: Set[Type[Command]] = set()
+    executed_commands: Set[Command] = set()
 
     for next_command_name in commands_passed:
 
-        next_command: Type[Command] = available_commands[next_command_name]
+        next_command: Command = available_commands[next_command_name]
 
         for prerequisite_command in next_command.get_prerequisites():
 
@@ -437,7 +448,7 @@ def quote_path(path: str) -> str:
         return path
 
 
-def get_package_version() -> str:
+def get_package_version() -> pkg_version.Version:
     """
     Retrieve the package version for the project from __init__ or _version
     """
@@ -459,7 +470,7 @@ def get_package_version() -> str:
     # noinspection PyUnresolvedReferences
     spec.loader.exec_module(version_module)
     # noinspection PyUnresolvedReferences
-    return version_module.__version__
+    return pkg_version.parse(version_module.__version__)
 
 
 def is_azure_build() -> bool:
@@ -469,13 +480,13 @@ def is_azure_build() -> bool:
     return "BUILD_REASON" in os.environ
 
 
-def version_string_to_url(version: str) -> str:
+def version_string_to_url(version: pkg_version.Version) -> str:
     """
     Make a Python package version string safe for URLs/folders.
 
     Our convention is to only replace all dots with dashes.
     """
-    return version.replace(".", "-")
+    return str(version).replace(".", "-")
 
 
 def check_sphinx_version() -> None:
@@ -499,15 +510,15 @@ Available program arguments:
     print(usage)
 
 
-available_commands: Dict[str, Type[Command]] = {
-    cmd.get_name(): cmd
+available_commands: Dict[str, Command] = {
+    cmd.name: cmd
     for cmd in (
-        Clean,
-        ApiDoc,
-        GettingStartedDoc,
-        Html,
-        Help,
-        FetchPkgVersions,
-        PrepareDocsDeployment,
+        Clean(),
+        ApiDoc(),
+        GettingStartedDoc(),
+        Html(),
+        Help(),
+        FetchPkgVersions(),
+        PrepareDocsDeployment(),
     )
 }
